Move explorer stories under live demos
