+++ stories/exampleTypes/ResultTable.stories.js
@@ -1,7 +1,7 @@
 import _ from 'lodash/fp'
 import React from 'react'
 import { storiesOf } from '@storybook/react'
-import TestTree from '../testTree'
+import TestTree from './testTree'
 import { ExampleTypes } from '../DemoControls'
 let { ResultTable } = ExampleTypes
