+++ stories/explorer/index.stories.js
@@ -1,6 +1,6 @@
 import 'babel-polyfill'
 import { storiesOf } from '@storybook/react'
 
-storiesOf('Components (Unthemed)|Index Explorer', module)
+storiesOf('Live Demos|Index Explorer', module)
   .addWithJSX('Advanced Search', require('./advanced').default)
   .addWithJSX('Basic Search', require('./basic').default)
